+++ molgenis-data-import/src/main/java/org/molgenis/data/importer/EntityImportReport.java
@@ -1,4 +1,4 @@
-package org.molgenis.framework.db;
+package org.molgenis.data.importer;
 
 import com.google.common.collect.Lists;
