+++ molgenis-data-import/src/main/java/org/molgenis/data/importer/emx/EmxImportJob.java
@@ -2,8 +2,8 @@ package org.molgenis.data.importer.emx;
 
 import org.molgenis.data.DatabaseAction;
 import org.molgenis.data.RepositoryCollection;
+import org.molgenis.data.importer.EntityImportReport;
 import org.molgenis.data.importer.ParsedMetaData;
-import org.molgenis.framework.db.EntityImportReport;
 
 /**
  * Parameter object for the import job.
