+++ molgenis-data-import/src/main/java/org/molgenis/data/importer/emx/EmxImportService.java
@@ -4,14 +4,10 @@ import com.google.common.collect.ImmutableMap;
 import org.molgenis.data.DataService;
 import org.molgenis.data.DatabaseAction;
 import org.molgenis.data.RepositoryCollection;
-import org.molgenis.data.importer.ImportService;
-import org.molgenis.data.importer.MetaDataParser;
-import org.molgenis.data.importer.ParsedMetaData;
+import org.molgenis.data.importer.*;
 import org.molgenis.data.meta.MetaDataService;
 import org.molgenis.data.meta.model.EntityType;
 import org.molgenis.data.support.GenericImporterExtensions;
-import org.molgenis.framework.db.EntitiesValidationReport;
-import org.molgenis.framework.db.EntityImportReport;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
