+++ molgenis-data-import/src/main/java/org/molgenis/data/importer/emx/MetaDataChanges.java
@@ -2,8 +2,7 @@ package org.molgenis.data.importer.emx;
 
 import com.google.common.collect.ImmutableList;
 import org.molgenis.data.Entity;
-import org.molgenis.data.importer.emx.ImportWriter;
-import org.molgenis.framework.db.EntityImportReport;
+import org.molgenis.data.importer.EntityImportReport;
 
 import java.util.ArrayList;
 import java.util.List;
