+++ molgenis-data-vcf/pom.xml
@@ -16,6 +16,11 @@
         </dependency>
         <dependency>
             <groupId>org.molgenis</groupId>
+            <artifactId>molgenis-data-import</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.molgenis</groupId>
             <artifactId>molgenis-data-validation</artifactId>
             <version>${project.version}</version>
         </dependency>
