+++ molgenis-file-ingester/pom.xml
@@ -26,6 +26,11 @@
         </dependency>
         <dependency>
             <groupId>org.molgenis</groupId>
+            <artifactId>molgenis-data-import</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.molgenis</groupId>
             <artifactId>molgenis-file</artifactId>
             <version>${project.version}</version>
         </dependency>
