+++ molgenis-ontology/src/main/java/org/molgenis/ontology/importer/OntologyImportService.java
@@ -4,14 +4,14 @@ import com.google.common.collect.Lists;
 import org.apache.commons.io.IOUtils;
 import org.molgenis.data.*;
 import org.molgenis.data.elasticsearch.SearchService;
+import org.molgenis.data.importer.EntitiesValidationReport;
 import org.molgenis.data.importer.EntitiesValidationReportImpl;
+import org.molgenis.data.importer.EntityImportReport;
 import org.molgenis.data.importer.ImportService;
 import org.molgenis.data.meta.MetaDataService;
 import org.molgenis.data.meta.model.EntityType;
 import org.molgenis.data.support.GenericImporterExtensions;
 import org.molgenis.data.support.QueryImpl;
-import org.molgenis.framework.db.EntitiesValidationReport;
-import org.molgenis.framework.db.EntityImportReport;
 import org.molgenis.ontology.core.meta.OntologyMetaData;
 import org.molgenis.security.permission.PermissionSystemService;
 import org.springframework.beans.factory.annotation.Autowired;
